drm:sun4i:hdmi: remove clk flag CLK_SET_RATE_PARENT
This is a workaround for the deeper issues surrounding the logic of the
clock handling of the display engines.

Uboot, with hdmi enabled, always enables clock halving for the TMDS
pads. The tmds clock is always a 2x clock of the two video parent
clocks. If halving is not set per default, the code used to go and halve
the top level parent clock. Resulting in everything being half the
frequency it is supposed to be, even the dotclock, on both pipelines.

By disabling this flag, the clock code will go and provide a useful
parent rate, and a useful target rate to the set_rate() callback,
without reclocking everything else first. This, to me, seems more
acceptable behaviour, especially since one usually wants the hdmi tmds
clock to match the dotclock.

Signed-off-by: Luc Verhaegen <libv@skynet.be>