kms:sun4i: split frontend fb setting
By making 3 different functions setting all of the relevant buffer
data, we simplify the code massively. So much so, that we can get away
with making it a big switch based on the format of the fb.

This big switch statement makes the whole thing immensely more readable
and logical, easy to debug and to extend.

Plus, we no longer confuse pitches, and we properly null values when we
switch number of planes or between tiling and linear.

Signed-off-by: Luc Verhaegen <libv@skynet.be>