kms:sunxi: add engine->crtc_mode_set() hook
This exposes both shortcomings in universal planes, and the superfluous
intermediate structure of _engine.

Our backends are half the conceptual old-skool style CRTC, and half the
sequencer/composer. It does not need to have any layer enabled or any
buffer attached to do something useful, it can just show a background
colour. Our 4 main layers are fully independent and can be ordered
(almost) at will. There is no primary layer.

Universal planes still asume that there is a primary layer. And also do
limited checking on whether this actually true.

The sun4i driver has, for some reason, an _engine struct in between,
that serves no purpose other than to provide separation where no
separation is needed, and it is clear in several cases that this
separation has holes in it. It also is a psychological barrier for
exposing things as the hardware sees it.

So instead of quickly bolting up some code to set the crtc width and
height in the backend, it was deferred to setting the primary plane.

So while KMS assumes that there is a primary plane, it never enforces
the primary plane to be the same dimensions as the CRTC.

So one is able to set the faux primary plane to whatever dimensions one
likes, which then has the backend set to those dimensions, which then
trips up our whole setup...

A very avoidable problem, on multiple fronts.

Also, we set interlacing with _every_ plane format set. This while this
conceptually belongs to the CRTC. The fact that the information comes
from the modeline should've tipped people off that this was wrong.

Let's at least get rid of this engine struct in future, and have direct
code dependencies between modules, so that at least this psychological
barrier disappears.

Signed-off-by: Luc Verhaegen <libv@skynet.be>