kms:sun4i: introduce frontend and yuv specific layers
All standard layers can be moved around as needed anyway, so we
can just tie one layer down to use the frontend (if needed), one layer
to use the backend specific yuv block, and two (a)rgb layers.

This significantly reduces complexity, as we can rely on a layer specific
atomic_check() function and layer specific formats to lock this down for us.

This does not fix the fact that KMS does not allow one to specify whether
layers scale or not. Userspace will just need to run into that wall. But
the likelyhood of a desktop application running into that is limited as
mostly media players will use scaling, and the frontend is deliberately
chosen to be the second layer, the first to support yuv.

It also does not fix the fact that we can only combine alpha in weird
ways. Perhaps we should make the frontend and the top level rgb layers
alpha only by giving only them an alpha property and the alpha formats?

Signed-off-by: Luc Verhaegen <libv@skynet.be>