dt:sun7i:bpi: hardwire parent and frequency for front and backend display
Our broken display clock setup gets in the way of driving our <7MHz lcd
and full HD on hdmi. It ties all front and backends to the same parent
clock for our lcd, which happens to be lower than the 148.5MHz of full
hd. Which then means that pixels are not delivered to the hdmi block on
time.

BPI equivalent of the lime2 fosdem fix.

Signed-off-by: Luc Verhaegen <libv@skynet.be>