kms:sunxi:backend: fix lowest layer alpha atomic check
Per-layer alpha works (an attribute), per-pixel alpha (an alpha component to
every pixel value), supposedly, does not on the lowest layer on sun[47]i.

The commit that was supposed to make this happen explained the above, but
made the exact opposite code change, stopping per-layer alpha, allowing
per-pixel alpha.

Was this ever tested or verified?

Fixes: e359c70462d2 ("drm/sun4i: backend: Support alpha property on lowest plane")
Signed-off-by: Luc Verhaegen <libv@skynet.be>